requirements.txt
# Web framework
fastapi>=0.118.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# Supabase & environment
supabase>=2.20.0
python-dotenv>=1.1.1

# Google AI/ML services
google-generativeai>=0.8.5
google-auth>=2.40.3
google-auth-oauthlib>=1.2.2
google-auth-httplib2>=0.2.0
google-api-python-client>=2.183.0
google-cloud-aiplatform>=1.117.0  # compatible with Python 3.11

# HTTP requests
requests>=2.32.0
charset-normalizer>=3.4.3

# Scientific computing & audio processing
numpy>=1.26.0
scipy>=1.12.0
librosa==0.10.1
audioread>=3.0.1
setuptools>=80.9.0
wheel>=0.45.1

# Data validation & models
pydantic>=2.11.9

# ML / AI - CPU only (Python 3.11 compatible)
torch>=2.7.1 --extra-index-url https://download.pytorch.org/whl/cpu
torchaudio>=2.7.1 --extra-index-url https://download.pytorch.org/whl/cpu

transformers==4.41.2
sentence-transformers==3.0.1

# Additional AI/ML dependencies
scikit-learn>=1.3.0
tokenizers>=0.14.3
huggingface-hub>=0.17.1
